Using regular expression to contraint the reserve memory
Before this commit, we just chage the reserve memory to the minimu
value that kdump needed if user enter a value too big or too small.
In this commit, if user enter a value too big or too small, he will
be prompted to enter again untill the value is in the reasonable range.