Pykickstart 3 support
Pykickstart 3 uses Argparse as the option parser, and there is
the new requirement for specifying when the given command has been
introduced, which can be used for kickstart file validation and/or
to generate documentation.

For more information about Pykickstart 2 to 3 porting see:
https://github.com/rhinstaller/pykickstart/blob/master/docs/2to3

So change the Kdump addon kickstart parsing code accordingly.